This one good question
The Tower of Hanoi is a classic puzzle that involves moving a tower of disks of different sizes from one peg to another. The objective of the puzzle is to move the entire tower to another peg, obeying the following simple rules:

Only one disk can be moved at a time.
Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack or on an empty peg.
No disk may be placed on top of a smaller disk.